backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sse-starlette>=2.1.0
python-dotenv>=1.0.0
langfuse>=3.0.0
